fix(save): container encounter global keys are unsigned
SaveGameExtractor.ParseContainer read the encounter record's two global keys
with ReadInt16, and the model held them as short. Both are wrong: the key space
runs past 32767.

GetGlobalValue @0x42250 compares its key with jb/jnb — UNSIGNED branches — and
its top band is the 56000+ range backed by global_flags2[]. A signed read turns
key 56012 into -9524, which then matches no band at all, so the encounter's gate
silently reads the wrong global.

Not hypothetical: the shipped OBJFIXED.DAT carries one such record (56252) and
the save games carry more — the committed dumps show -9524 and -9221, i.e. 56012
and 56315. Nothing consumes the field at runtime yet, so this is latent rather
than a live misbehaviour, but it is baked into the model type and would bite
whoever wires handle_Catapult / handle_RiftMachine (TASK-136), which are the two
functions that read it.

Held as int rather than ushort so the value reads as the number the engine uses.

FOUND WHILE SIZING TASK-162, which frames the CLI's separate Container reader as
the redundant one to fold away. On this field the CLI is the one that is RIGHT
(it reads ReadUInt16) and the shared parser had the bug — so folding as-is would
have propagated it. The other three signedness differences between the two
readers (worldItemId, x/y, globalStateIndex) have no high-bit-set values in the
shipped data, so they are latent only.

4 tests.

+++ GameData/Resources/Data/SaveGameContainerEncounterData.cs
@@ -2,8 +2,8 @@ namespace GameData.Resources.Data;
 
 public class SaveGameContainerEncounterData {
     public SaveGameContainerEncounterData(
-        short globalDataKey1,
-        short globalDataKey2,
+        int globalDataKey1,
+        int globalDataKey2,
         byte gdsNumber,
         byte gdsLetter,
         byte firesTrapEncounter,
@@ -19,8 +19,23 @@ public class SaveGameContainerEncounterData {
         Y = y;
     }
 
-    public short GlobalDataKey1 { get; }
-    public short GlobalDataKey2 { get; }
+    /// <summary>
+    /// The global whose value gates this encounter — a key for <c>GetGlobalValue</c> @0x42250.
+    /// </summary>
+    /// <remarks>
+    /// <b>Unsigned, and it must be: the key space runs past 32767.</b> <c>GetGlobalValue</c>
+    /// compares the key with <c>jb</c>/<c>jnb</c> — unsigned branches — and its top band is the
+    /// 56000+ range backed by <c>global_flags2[]</c>. Reading the field as a signed 16-bit turns
+    /// key 56012 into -9524, which no band then matches.
+    ///
+    /// <para>This is not hypothetical: the shipped OBJFIXED.DAT has one such record and the save
+    /// games have more (56012 and 56315 both appear). Held as <c>int</c> rather than
+    /// <c>ushort</c> so the value reads as the number the engine uses.</para>
+    /// </remarks>
+    public int GlobalDataKey1 { get; }
+
+    /// <inheritdoc cref="GlobalDataKey1"/>
+    public int GlobalDataKey2 { get; }
     public byte GdsNumber { get; }
     public byte GdsLetter { get; }
